requirements.txt
selenium
webdriver-manager
beautifulsoup4
PyYAML
